[update]still fail to cope with duplicate

+++ src/kvraft/common.go
@@ -16,7 +16,7 @@ type PutAppendArgs struct {
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
-	ClientId int
+	ClientId int64
 	RequestId int
 }
